fix build and checkstyle
