+++ Task_1_1_3/build.gradle
@@ -3,7 +3,7 @@ plugins {
     id 'jacoco'
 }
 
-group = 'ru.nsu.syspro.zagitov.operations_with_equations'
+group = 'ru.nsu.syspro.zagitov.operationswithequations'
 version = '1.0-SNAPSHOT'
 
 repositories {
