+++ Task_1_1_3/src/main/java/ru/nsu/syspro/zagitov/operationswithequations/Sub.java
@@ -1,4 +1,4 @@
-package ru.nsu.syspro.zagitov.operations_with_equations;
+package ru.nsu.syspro.zagitov.operationswithequations;
 
 import java.util.ArrayList;
